factor out DealQuery
